+++ src/hivememory/engines/memory_compiler/__init__.py
@@ -1,8 +1,4 @@
-"""
-MemoryCompiler — 统一的记忆到文本编译层。
-
-Phase 1: 作为现有 MemoryAtomRenderer 和 PendingAtomRenderer 的外观层。
-"""
+"""MemoryCompiler unified memory-to-artifact compilation layer."""
 
 from hivememory.engines.memory_compiler.compiler import MemoryCompiler
 from hivememory.engines.memory_compiler.models import (
