+++ src/hivememory/engines/memory_compiler/adapters.py
@@ -1,17 +1,20 @@
-"""薄适配器层 — 委托给现有 MemoryAtomRenderer。"""
+"""Compatibility adapter functions for MemoryAtom compilation."""
 
 from __future__ import annotations
 
 from hivememory.core.models import MemoryAtom
-from hivememory.utils.memory_atom_renderer import MemoryAtomRenderer
+from hivememory.engines.memory_compiler.handlers.memory_atom import (
+    _render_agent_profile,
+    _render_dense_embedding,
+    _render_full_context,
+    _render_index_context,
+    _render_sparse_embedding,
+)
 from hivememory.engines.memory_compiler.models import MemoryCompileOptions
 
 
-# ========== MemoryAtom 适配器 ==========
-
-
 def compile_atom_full(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
-    return MemoryAtomRenderer.for_full_context(
+    return _render_full_context(
         memory=atom,
         max_content_length=options.max_content_length,
         stale_days=options.stale_days,
@@ -19,7 +22,7 @@ def compile_atom_full(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
 
 
 def compile_atom_index(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
-    return MemoryAtomRenderer.for_index_context(
+    return _render_index_context(
         memory=atom,
         max_summary_length=options.max_summary_length,
         stale_days=options.stale_days,
@@ -27,12 +30,12 @@ def compile_atom_index(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
 
 
 def compile_atom_dense(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
-    return MemoryAtomRenderer.for_dense_embedding(atom)
+    return _render_dense_embedding(atom)
 
 
 def compile_atom_sparse(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
-    return MemoryAtomRenderer.for_sparse_embedding(atom)
+    return _render_sparse_embedding(atom)
 
 
 def compile_atom_agent_profile(atom: MemoryAtom, options: MemoryCompileOptions) -> str:
-    return MemoryAtomRenderer.for_agent_profile(atom)
+    return _render_agent_profile(atom)
